dom functions moved to Dom class + edit game UI finished
