+++ package.json
@@ -59,6 +59,7 @@
     "jquery": "^3.2.1",
     "jquery-contextmenu": "^2.5.0",
     "mkdirp": "^0.5.1",
+    "moment": "^2.18.1",
     "rimraf": "^2.6.1",
     "sass-percentage-circle": "^0.2.0",
     "truncate": "^2.0.0",
@@ -74,6 +75,7 @@
     "@types/jquery.contextmenu": "^1.7.33",
     "@types/mkdirp": "^0.5.1",
     "@types/mocha": "^2.2.42",
+    "@types/moment": "^2.13.0",
     "@types/node": "^8.0.25",
     "@types/rimraf": "0.0.28",
     "@types/uuid": "^3.4.1",
